Fixed minor bugs around Default Core determination
Items in Kits were having their default core replaced, then immediately getting it added back in. This change ensures that the Default Core is only affected when adding the item independently of a Kit, and it defines the behavior when an item is not the default for a particular slot.