Added functions to import Emblem Palette Images
These functions include the ability to check the ETag of an image and avoid pulling the image itself down if it has not been modified. If it has been modified, then the ETag is used to recognize that and delete the old version before pulling the new version. These functions also split out the Emblem Palettes folder from the normal folder dict and move it to a special Emblem Palette folder dict, which is needed due to the high volume of images involved.