Added Emblem Palette File Constants
The FOLDERS_LIMIT sets how many folders can be returned in a single file-system query. Currently, this is set to 500 since there are 299 active Emblem Palettes, but this may need to be increased or future-proofed eventually. The EMBLEM_PALETTE_ROOT_FOLDER config indicates the base folder for Emblem Palettes within the Customization Images folder.