Fixed an issue with default cross-core items
When Visors became cross-core on Tuesday, some of them flagged a change to their Default of Core field. The issue was due to the fact that we skip processing for items that show up repeatedly, which is true of all cross-core items, and some core-specific items (specifically Kit items). We added some logic to ensure that items marked as defaults for a core are reprocessed, which will add the necessary Default of Core value to the item when relevant.