Added ETag functionality to getCustomizationItem()
We need to start retrieving ETags when pulling data from the Waypoint API, to help avoid pulling the same data again later. This addition support backwards compatibility (setting the third argument to false or not including it will still just return the original JSON), but it also adds the ETag into the return if requested.