Added additional constant definitions
The global customization search required the addition of a SearchCategory constant to each of the customization categories. I also began moving some of the customizations into the ArmorCustomization file for the Armor section, rather than defining them all in the CustomizationConstants file. This work is purely topical, and has no impact on the functionality of any existing code.