Added QuickSearch Support
QuickSearch limits the number of items returned by a query and drastically reduces the processing required to expedite the process of getting just the names of the relevant items and getting out. It performs a "startsWith" query instead of a "contains" query to better match the user's query with an actual item, but this can be easily changed.