Improved Twitch Drop Reward Structure
The Twitch Drop Rewards have now been moved to their own database, with a reference linking the existing Twitch Drop database to the rewards database. This will drastically reduce the amount of manual intervention needed to properly send out notifications and makes it possible to create a Twitch Drops list and page. More to come.