Reworked ETag usage to abort processing much earlier
ETags are now compared when an individual item is up for processing. We can't skip attachments or kit items this early (or anything marked as force check), but we can skip a bunch of items that don't need to process.

I also added the ability to get the ETag directly from the API when the guide/xo endpoint does not have it. This requires slightly more effort to check, but it is still much faster than processing the entire item because we don't have an ETag.